+++ R/git-utils.R
@@ -260,9 +260,8 @@ check_branch <- function(branch) {
 
 check_branch_pulled <- function(branch = git_branch_name(), use = "git pull") {
   remote <- git_branch_tracking_FIXME(branch)
-  # if no remote branch, nothing to check
   if (is.null(remote)) {
-    ui_done("No remote branch yet for local branch {ui_value(branch)}")
+    ui_done("Local branch {ui_value(branch)} is not tracking a remote branch.")
     return(invisible())
   }